coverity issues

+++ util/lpcx_main.c
@@ -98,7 +98,7 @@ int main(int argc, char **argv)
     if (hdr.npoles+hdr.nvals > 0
         && hdr.npoles+hdr.nvals < 0x0FFFFFFF
         && hdr.npoles > 0) {
-      coef = (MYFLT *)malloc((hdr.npoles+hdr.nvals)*sizeof(MYFLT));
+      coef = (MYFLT *)malloc(((uint64_t)hdr.npoles+hdr.nvals)*sizeof(MYFLT));
       for (i = 0; i<floor(hdr.framrate*hdr.duration); i++) {
         if (UNLIKELY(fread(coef, sizeof(MYFLT), hdr.npoles,inf) != hdr.npoles)) {
           fprintf(stderr, Str("Read failure\n"));
